README.md
     A project that I created in my journey of learning JAVA development in JetBrains Academy.
During this course, I was able to gain a lot of knowledge in generics, and different kinds of Trees in JAVA. I have used the Factory design pattern to create different kinds of finders. I have also used different kinds of sorting algorithms such as merge sort and quick Sort.